working version corrected
